more UI changes, remove old code, move things out of status bar, code review changes / fixes

+++ data/sql/version-001.sql
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS qsos (
   park_ref TEXT NOT NULL,
   callsign TEXT NOT NULL,
   mode TEXT,
-  frequency_khz INTEGER,
+  frequency_khz REAL,
   created_utc DATETIME NOT NULL,
   spotter TEXT,
   spotter_comment TEXT,
